% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistantkernel-deprecated.R
\name{resistantkernel-deprecated}
\alias{resistantkernel-deprecated}
\alias{raw.spread}
\title{Deprecated function(s) in the resistantkernel package}
\usage{
raw.spread(x, spread.value, row, col)
}
\arguments{
\item{x, row, col}{Passed as is to \link{raw_spread}}

\item{spread.value}{passed as \code{spread_value} to \code{\link[=raw_spread]{raw_spread()}}}

\item{...}{Parameters to be passed to the current version of the functions}
}
\value{
See current versions of functions.
}
\description{
These functions are provided for compatibility with older version of
the resistantkernel and the earlier \pkg{spread} and \pkg{gridprocess}
packages.  They may eventually be completely removed.

\code{raw.spread()} has been replaced by \code{\link[=raw_spread]{raw_spread()}}
}
\keyword{internal}
