% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistantkernel-package.R
\docType{package}
\name{resistantkernel-package}
\alias{resistantkernel-package}
\alias{_PACKAGE}
\alias{resistantkernel}
\title{resistantkernel: Calculates least cost paths and resistant kernels}
\description{
These functions are useful for measuring (least cost) functional distances based on a resistant surface, and the size of a resistant kernel for points on that surface (one way of measuring habitat quality).
}
\keyword{internal}
